Do not include cpl_float.h in gdal_priv.h, to avoid issues on Windows with min/max macros
Fixes #12338
