+++ gcore/gdal_priv.h
@@ -47,7 +47,6 @@ class GDALAlgorithm;
 #include "gdalsubdatasetinfo.h"
 #include "cpl_vsi.h"
 #include "cpl_conv.h"
-#include "cpl_float.h"
 #include "cpl_string.h"
 #include "cpl_minixml.h"
 #include "cpl_multiproc.h"
@@ -3748,10 +3747,10 @@ class CPL_DLL GDALMDArray : virtual public GDALAbstractMDArray,
     Transpose(const std::vector<int> &anMapNewAxisToOldAxis) const;
 
     std::shared_ptr<GDALMDArray> GetUnscaled(
-        double dfOverriddenScale = cpl::NumericLimits<double>::quiet_NaN(),
-        double dfOverriddenOffset = cpl::NumericLimits<double>::quiet_NaN(),
+        double dfOverriddenScale = std::numeric_limits<double>::quiet_NaN(),
+        double dfOverriddenOffset = std::numeric_limits<double>::quiet_NaN(),
         double dfOverriddenDstNodata =
-            cpl::NumericLimits<double>::quiet_NaN()) const;
+            std::numeric_limits<double>::quiet_NaN()) const;
 
     virtual std::shared_ptr<GDALMDArray>
     GetMask(CSLConstList papszOptions) const;
@@ -4627,29 +4626,21 @@ GDALDataset *GDALCreateOverviewDataset(GDALDataset *poDS, int nOvrLevel,
 // TODO: The expression `abs(fVal1 + fVal2)` looks strange; is this a bug?
 // Should this be `abs(fVal1) + abs(fVal2)` instead?
 
-inline bool ARE_REAL_EQUAL(GFloat16 dfVal1, GFloat16 dfVal2, int ulp = 2)
-{
-    using std::abs;
-    return dfVal1 == dfVal2 || /* Should cover infinity */
-           abs(dfVal1 - dfVal2) < cpl::NumericLimits<GFloat16>::epsilon() *
-                                      abs(dfVal1 + dfVal2) * ulp;
-}
-
 inline bool ARE_REAL_EQUAL(float fVal1, float fVal2, int ulp = 2)
 {
     using std::abs;
     return fVal1 == fVal2 || /* Should cover infinity */
            abs(fVal1 - fVal2) <
-               cpl::NumericLimits<float>::epsilon() * abs(fVal1 + fVal2) * ulp;
+               std::numeric_limits<float>::epsilon() * abs(fVal1 + fVal2) * ulp;
 }
 
-// We are using `cpl::NumericLimits<float>::epsilon()` for backward
+// We are using `std::numeric_limits<float>::epsilon()` for backward
 // compatibility
 inline bool ARE_REAL_EQUAL(double dfVal1, double dfVal2, int ulp = 2)
 {
     using std::abs;
     return dfVal1 == dfVal2 || /* Should cover infinity */
-           abs(dfVal1 - dfVal2) < cpl::NumericLimits<float>::epsilon() *
+           abs(dfVal1 - dfVal2) < std::numeric_limits<float>::epsilon() *
                                       abs(dfVal1 + dfVal2) * ulp;
 }
